pyproject.toml
[tool.poetry]
name = "x-developper-challenge-search"
version = "0.1.0"
description = ""
authors = ["github-actions <jojoto8845@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
gensim = "^4.3.2"
scipy = "1.10.1"
scikit-learn = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
